Move the conversion test into the legacy module
